Remove Isaac Guice module and merge Isaac and Segue config
Merging the two Guice configuration modules requires some work to ensure
that singletons in the only-used-once IsaacGuiceConfigurationModule do not
get initialised multiple times by the created-many-times Segue Guice
config.
I removed some unused Segue configuration and standalone files too, since
we do not use Segue separately and do not plan to.
I also merged the ContentMapper reflections scans to scan both Isaac and
Segue classes in one go, since this scan is the one that has caused Jackson
and Orika to map things incorrectly or fail to map under heavy load at
startup time.
