+++ src/main/java/uk/ac/cam/cl/dtg/segue/configuration/SegueGuiceConfigurationModule.java
@@ -33,11 +33,29 @@ import org.elasticsearch.client.Client;
 import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.ac.cam.cl.dtg.isaac.api.managers.AssignmentManager;
 import uk.ac.cam.cl.dtg.isaac.api.managers.GameManager;
+import uk.ac.cam.cl.dtg.isaac.api.managers.QuizAssignmentManager;
+import uk.ac.cam.cl.dtg.isaac.api.managers.URIManager;
+import uk.ac.cam.cl.dtg.isaac.api.services.EmailService;
+import uk.ac.cam.cl.dtg.isaac.api.services.GroupChangedService;
+import uk.ac.cam.cl.dtg.isaac.dao.GameboardPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.IAssignmentPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.IQuizAssignmentPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.IQuizAttemptPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.IQuizQuestionAttemptPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.PgAssignmentPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.PgQuizAssignmentPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.PgQuizAttemptPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.dao.PgQuizQuestionAttemptPersistenceManager;
+import uk.ac.cam.cl.dtg.isaac.quiz.IsaacSymbolicChemistryValidator;
+import uk.ac.cam.cl.dtg.isaac.quiz.IsaacSymbolicLogicValidator;
+import uk.ac.cam.cl.dtg.isaac.quiz.IsaacSymbolicValidator;
 import uk.ac.cam.cl.dtg.segue.api.Constants;
 import uk.ac.cam.cl.dtg.segue.api.managers.ExternalAccountManager;
 import uk.ac.cam.cl.dtg.segue.api.managers.GroupManager;
 import uk.ac.cam.cl.dtg.segue.api.managers.IExternalAccountManager;
+import uk.ac.cam.cl.dtg.segue.api.managers.IGroupObserver;
 import uk.ac.cam.cl.dtg.segue.api.managers.IStatisticsManager;
 import uk.ac.cam.cl.dtg.segue.api.managers.ITransactionManager;
 import uk.ac.cam.cl.dtg.segue.api.managers.IUserAccountManager;
@@ -144,21 +162,20 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
     private static IQuestionAttemptManager questionPersistenceManager = null;
     private static SegueJobService segueJobService = null;
 
-    //private static ILogManager logManager;
     private static LogManagerEventPublisher logManager;
     private static EmailManager emailCommunicationQueue = null;
     private static IMisuseMonitor misuseMonitor = null;
     private static IMetricsExporter metricsExporter = null;
     private static StatisticsManager statsManager = null;
-    //private static IStatisticsManager statsManager = null;
     private static GroupManager groupManager = null;
-    private static IUserAlerts userAlerts = null;
-    private static IUserStreaksManager userStreaksManager = null;
-    private static IUserBadgePersistenceManager userBadgePersitenceManager = null;
     private static IExternalAccountManager externalAccountManager = null;
+    private static GameboardPersistenceManager gameboardPersistenceManager = null;
+    private static SchoolListReader schoolListReader = null;
+    private static AssignmentManager assignmentManager = null;
+    private static IGroupObserver groupObserver = null;
 
     private static Collection<Class<? extends ServletContextListener>> contextListeners;
-    private static Map<String, Reflections> reflections = com.google.common.collect.Maps.newHashMap();
+    private static final Map<String, Reflections> reflections = com.google.common.collect.Maps.newHashMap();
 
     /**
      * Create a SegueGuiceConfigurationModule.
@@ -270,8 +287,12 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
                         .getProperty(Constants.REMOTE_GIT_SSH_KEY_PATH)));
 
         bind(IUserGroupPersistenceManager.class).to(PgUserGroupPersistenceManager.class);
-
         bind(IAssociationDataManager.class).to(PgAssociationDataManager.class);
+        bind(IAssignmentPersistenceManager.class).to(PgAssignmentPersistenceManager.class);
+        bind(IQuizAssignmentPersistenceManager.class).to(PgQuizAssignmentPersistenceManager.class);
+        bind(IQuizAttemptPersistenceManager.class).to(PgQuizAttemptPersistenceManager.class);
+        bind(IQuizQuestionAttemptPersistenceManager.class).to(PgQuizQuestionAttemptPersistenceManager.class);
+        bind(IUserBadgePersistenceManager.class).to(PgUserBadgePersistenceManager.class);
     }
 
     /**
@@ -336,6 +357,8 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
         bind(AbstractUserPreferenceManager.class).to(PgUserPreferenceManager.class);
 
         bind(IUserAlerts.class).to(PgUserAlerts.class);
+
+        bind(IUserStreaksManager.class).to(PgUserStreakManager.class);
 
         bind(IStatisticsManager.class).to(StatisticsManager.class);
 
@@ -488,7 +511,7 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
     @Singleton
     private static ContentMapper getContentMapper() {
         if (null == mapper) {
-            mapper = new ContentMapper(getReflectionsClass("uk.ac.cam.cl.dtg.segue"));
+            mapper = new ContentMapper(getReflectionsClass("uk.ac.cam.cl.dtg"));
             log.info("Creating Singleton of the Content Mapper");
         }
 
@@ -701,6 +724,19 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
         return groupManager;
     }
 
+
+    @Inject
+    @Provides
+    @Singleton
+    private IGroupObserver getGroupObserver(EmailManager emailManager, GroupManager groupManager, UserAccountManager userManager,
+                                            AssignmentManager assignmentManager, QuizAssignmentManager quizAssignmentManager) {
+        if (null == groupObserver) {
+            groupObserver = new GroupChangedService(emailManager, groupManager, userManager, assignmentManager, quizAssignmentManager);
+            log.info("Creating singleton of GroupObserver");
+        }
+        return groupObserver;
+    }
+
     /**
      * Get singleton of misuseMonitor.
      *
@@ -819,41 +855,6 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
     }
 
     /**
-     * Gets instance of user badge database liason manager
-     *
-     * @param postgresDB database
-     * @return concrete instance of IUserBadgePersistenceManager
-     */
-    @Provides
-    @Singleton
-    @Inject
-    private static IUserBadgePersistenceManager getUserBadgePersistenceManager(final PostgresSqlDb postgresDB) {
-
-        if (null == userBadgePersitenceManager) {
-            userBadgePersitenceManager = new PgUserBadgePersistenceManager(postgresDB);
-        }
-        return userBadgePersitenceManager;
-    }
-
-
-    /**
-     * Gets instance of the user streaks manager
-     *
-     * @param postgresDB database
-     * @return concrete instance of IUserStreaksManager
-     */
-    @Provides
-    @Singleton
-    @Inject
-    private static IUserStreaksManager getUserStreaksManager(final PostgresSqlDb postgresDB) {
-
-        if (null == userStreaksManager) {
-            userStreaksManager = new PgUserStreakManager(postgresDB);
-        }
-        return userStreaksManager;
-    }
-
-    /**
      * Gets the instance of the StatisticsManager. Note: this class is a hack and needs to be refactored.... It is
      * currently only a singleton as it keeps a cache.
      *
@@ -960,6 +961,131 @@ public class SegueGuiceConfigurationModule extends AbstractModule implements Ser
     @Provides
     private IPLocationResolver getIPLocator(@Named(Constants.IP_INFO_DB_API_KEY) final String apiKey) {
         return new IPInfoDBLocationResolver(apiKey);
+    }
+
+    /**
+     * Gets a Game persistence manager.
+     *
+     * This needs to be a singleton as it maintains temporary boards in memory.
+     *
+     * @param database
+     *            - the database that persists gameboards.
+     * @param contentManager
+     *            - api that the game manager can use for content resolution.
+     * @param mapper
+     *            - an instance of an auto mapper for translating gameboard DOs and DTOs efficiently.
+     * @param objectMapper
+     *            - a mapper to allow content to be resolved.
+     * @param uriManager
+     *            - so that the we can create content that is aware of its own location
+     * @return Game persistence manager object.
+     */
+    @Inject
+    @Provides
+    @Singleton
+    private static GameboardPersistenceManager getGameboardPersistenceManager(final PostgresSqlDb database,
+                                                                              final IContentManager contentManager, final MapperFacade mapper, final ObjectMapper objectMapper,
+                                                                              final URIManager uriManager, @Named(CONTENT_INDEX) final String contentIndex) {
+        if (null == gameboardPersistenceManager) {
+            gameboardPersistenceManager = new GameboardPersistenceManager(database, contentManager, mapper,
+                    objectMapper, uriManager, contentIndex);
+            log.info("Creating Singleton of GameboardPersistenceManager");
+        }
+
+        return gameboardPersistenceManager;
+    }
+
+    /**
+     * Gets an assignment manager.
+     *
+     * This needs to be a singleton because operations like emailing are run for each IGroupObserver, the
+     * assignment manager should only be one observer.
+     *
+     * @param assignmentPersistenceManager
+     *            - to save assignments
+     * @param groupManager
+     *            - to allow communication with the group manager.
+     * @param emailService
+     *            - email service
+     * @param gameManager
+     *            - the game manager object
+     * @param properties
+     *            - properties loader for the service's hostname
+     * @return Assignment manager object.
+     */
+    @Inject
+    @Provides
+    @Singleton
+    private static AssignmentManager getAssignmentManager(
+            final IAssignmentPersistenceManager assignmentPersistenceManager, final GroupManager groupManager,
+            final EmailService emailService, final GameManager gameManager, final PropertiesLoader properties) {
+        if (null == assignmentManager) {
+            assignmentManager =  new AssignmentManager(assignmentPersistenceManager, groupManager, emailService, gameManager, properties);
+            log.info("Creating Singleton AssignmentManager");
+        }
+        return assignmentManager;
+    }
+
+    /**
+     * Gets an instance of the symbolic question validator.
+     *
+     * @return IsaacSymbolicValidator preconfigured to work with the specified checker.
+     */
+    @Provides
+    @Singleton
+    @Inject
+    private static IsaacSymbolicValidator getSymbolicValidator(PropertiesLoader properties) {
+
+        return new IsaacSymbolicValidator(properties.getProperty(Constants.EQUALITY_CHECKER_HOST),
+                properties.getProperty(Constants.EQUALITY_CHECKER_PORT));
+    }
+
+    /**
+     * Gets an instance of the chemistry question validator.
+     *
+     * @return IsaacSymbolicChemistryValidator preconfigured to work with the specified checker.
+     */
+    @Provides
+    @Singleton
+    @Inject
+    private static IsaacSymbolicChemistryValidator getSymbolicChemistryValidator(PropertiesLoader properties) {
+
+        return new IsaacSymbolicChemistryValidator(properties.getProperty(Constants.CHEMISTRY_CHECKER_HOST),
+                properties.getProperty(Constants.CHEMISTRY_CHECKER_PORT));
+    }
+
+    /**
+     * Gets an instance of the symbolic logic question validator.
+     *
+     * @return IsaacSymbolicLogicValidator preconfigured to work with the specified checker.
+     */
+    @Provides
+    @Singleton
+    @Inject
+    private static IsaacSymbolicLogicValidator getSymbolicLogicValidator(PropertiesLoader properties) {
+
+        return new IsaacSymbolicLogicValidator(properties.getProperty(Constants.EQUALITY_CHECKER_HOST),
+                properties.getProperty(Constants.EQUALITY_CHECKER_PORT));
+    }
+
+    /**
+     * This provides a singleton of the SchoolListReader for use by segue backed applications..
+     *
+     * We want this to be a singleton as otherwise it may not be threadsafe for loading into same SearchProvider.
+     *
+     * @param provider
+     *            - The search provider.
+     * @return schoolList reader
+     */
+    @Inject
+    @Provides
+    @Singleton
+    private SchoolListReader getSchoolListReader(final ISearchProvider provider) {
+        if (null == schoolListReader) {
+            schoolListReader = new SchoolListReader(provider);
+            log.info("Creating singleton of SchoolListReader");
+        }
+        return schoolListReader;
     }
 
     /**
