+++ src/main/java/uk/ac/cam/cl/dtg/segue/dos/PgUserStreakManager.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
+import com.google.inject.Inject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.ac.cam.cl.dtg.segue.api.userAlerts.UserAlertsWebSocket;
@@ -32,6 +33,7 @@ public class PgUserStreakManager implements IUserStreaksManager {
      * @param database
      *            client for postgres.
      */
+    @Inject
     public PgUserStreakManager(final PostgresSqlDb database) {
         this.database = database;
     }
